Use `ipl\Web\Compat\CompatForm` instead of `Icinga\Module\Jira\Web\Form` as base form for NewIssueForm
Since Icinga\Module\Jira\Web\Form, library/Jira/Web/Form/LegacyDecorator, form-compat26 css class and changes
to form-element-description css class are removed.
