+++ public/css/module.less
@@ -72,59 +72,3 @@ img.icon {
   overflow: hidden;
   display: inline-block;
 }
-
-.form-element-description {
-    color: @text-color;
-    font-style: italic;
-    margin: 0.25em 0 0 15em;
-    padding-left: 0.5em;
-    width: 100%;
-}
-
-.form-compat26 {
-    .form-element-description {
-        max-width: 36em;
-        padding: 0;
-        width: auto;
-    }
-
-    .control-label-group {
-        width: 15em;
-    }
-
-    select {
-        max-width: 36em;
-        min-width: 20em;
-        line-height: 2em;
-        height: 2.4em;
-        padding-left: 0.5em;
-        border-style: solid;
-        border-color: transparent;
-        border-bottom-color: @gray-light;
-        border-width: 1px 1px 1px 3px;
-        background-color: @low-sat-blue;
-    }
-
-    select option {
-        height: 2em;
-        padding-top: 0.3em;
-    }
-
-    select[multiple] {
-        height: auto;
-    }
-
-    select[multiple=multiple] {
-        height: auto;
-    }
-
-    input[type=text], input[type=password], textarea, select {
-        max-width: 36em;
-        min-width: 20em;
-        width: 100%;
-    }
-
-    select:focus, textarea:focus {
-        border-color: @icinga-blue;
-    }
-}
